refactor: 패키지 이동 #12 /infrastructure -> /infrastructure/websocket
